% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_colors.R
\name{are_colors}
\alias{are_colors}
\title{Are Colors?}
\usage{
are_colors(color_vector)
}
\arguments{
\item{color_vector}{vector; A vector of color values.}
}
\value{
A logical vector.
}
\description{
Determines if the input is a vector of valid colors according
to the \code{grDevices::col2rgb} function.
}
\details{
Follows the advice from \href{https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}{SO}
}
