% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_model_barplot.R
\name{feature_model_barplot}
\alias{feature_model_barplot}
\title{Create a Feature By Ecological Model Barplot}
\usage{
feature_model_barplot(
  model_summary,
  habitat_zone,
  ylab,
  main,
  location = "topleft"
)
}
\arguments{
\item{model_summary}{data.frame; A data frame returned by the
\code{summary_table} function.}

\item{habitat_zone}{raster; A habitat zone raster returned by the
\code{set_habitat_zone} function.}

\item{ylab}{character; The plot y-axis label.}

\item{main}{character; The plot main label.}

\item{location}{character; \code{legend} location keyword. One of:
"bottomright", "bottom", "bottomleft", "left",
"topleft", "top", "topright", "right" or "center".}
}
\value{
A \code{barplot} object.
}
\description{
Creates a barplot from a model summary data frame.
}
