% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_colortable.R
\name{has_colortable}
\alias{has_colortable}
\title{Has Colortable?}
\usage{
has_colortable(raster)
}
\arguments{
\item{raster}{}
}
\value{
Returns TRUE if the raster has a valid colortable.
}
\description{
Determines if a RasterLayer has a valid colortable defined.
}
