% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSIcalc.R
\name{HSIcalc}
\alias{HSIcalc}
\title{Calculate Habitat Suitability Index (HSI)}
\usage{
HSIcalc(si_list, method = "mean")
}
\arguments{
\item{si_list}{list; A list of suitability indices.}

\item{method}{character; The method for calculating the suitability
index. Must be one of "mean",}
}
\value{
An HSI object matching the data format of the input suitability
indices.
}
\description{
Calculates a habitat suitability index for the input
suitability indices using the specified method.
}
