% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSIeqtn.R
\name{HSIeqtn}
\alias{HSIeqtn}
\title{Computes Habitat Suitability Index based on Model-Specified Equation}
\usage{
HSIeqtn(HSImodelname, SIV, HSImetadata)
}
\arguments{
\item{HSImodelname}{character; A character string in quotations that
must match an existing model name in HSImetadata.}

\item{SIV}{list; a vector of suitability index values used in
the model specified in HSImodelname.}

\item{HSImetadata}{data frame; A data frame of HSI model metadata within
the ecorest package.}
}
\value{
A numeric of the habitat suitability index ranging from 0 to 1.
}
\description{
HSIeqtn computes a habitat suitability index based on
equations specified in U.S. Fish and Wildlife Service habitat
suitability models contained within ecorest via HSImodels and HSImetadata.
Habitat suitability indices represent an overall assessment of habitat
quality from combining individual suitability indices for multiple
independent variables. The function computes an overall habitat
suitability index.
}
\references{
\itemize{
\item US Fish and Wildlife Service. (1980). Habitat as a basis for environmental
assessment. Ecological Services Manual, 101.
\item US Fish and Wildlife Service. (1980). Habitat Evaluation Procedures (HEP).
Ecological Services Manual, 102.
\item US Fish and Wildlife Service. (1981). Standards for the Development of
Habitat Suitability Index Models. Ecological Services Manual, 103.
}
}
