% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSIplotter.R
\name{HSIplotter}
\alias{HSIplotter}
\title{Plots habitat suitability index curves}
\usage{
HSIplotter(SI)
}
\arguments{
\item{SI}{matrix; matrix of suitability curves ordered as
parameter breakpoints and associated suitability
indices for each parameter with appropriate column
names.}
}
\value{
A multipanel figure displaying suitability curves.
}
\description{
Plots habitat suitability index curves for all metrics in
the model.
}
\examples{
# Get barredowl HSI model record from the `ecorest`package (example of
# continuous metrics)
barredowl_ecorest <- ecorest::HSImodels$barredowl

# Create the plot
HSIplotter(barredowl_ecorest)

# Get americancoot HSI model record from the `ecorest`package (example of
continuous and categorical metrics)
americancoot_ecorest <- ecorest::HSImodels$americancoot

# Create the plot
HSIplotter(americancoot_ecorest)

}
\references{
\itemize{
\item US Fish and Wildlife Service. (1980). Habitat as a basis for environmental
assessment. Ecological Services Manual, 101.
\item US Fish and Wildlife Service. (1980). Habitat Evaluation Procedures (HEP).
Ecological Services Manual, 102.
\item US Fish and Wildlife Service. (1981). Standards for the Development of
Habitat Suitability Index Models. Ecological Services Manual, 103.
}
}
