% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelse.R
\name{ifelse}
\alias{ifelse}
\title{Raster Con Function}
\usage{
ifelse(condition, true_value, false_value)
}
\arguments{
\item{condition}{raster or numeric; an expression that results in a
boolean, matching the type of the condition.}

\item{false_value}{}
}
\value{
the true or false value matching the data type of the condition.
}
\description{
This function is designed to behave similar to the ESRI
Spatial Analyst \code{Con} function. The purpose of this function is to
provide R with a raster safe \code{ifelse} function.
}
\details{
This function was proposed by
\href{https://gis.stackexchange.com/users/865/spacedman}{Spacedman} on
gis.stackexchange way back in 2013,
\href{https://gis.stackexchange.com/questions/69734/}{What is the equivalent of arcpy "Con" in QGIS and/or R raster-package?}
}
\examples{

}
