% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cat_raster.R
\name{is_cat_raster}
\alias{is_cat_raster}
\title{Is Catagorical Raster?}
\usage{
is_cat_raster(ras)
}
\arguments{
\item{ras}{raster; The raster layer to test.}
}
\value{
logical; TRUE if class is of type \code{raster::RasterLayer} and a
factor.
}
\description{
Is Catagorical Raster?
}
