% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_RasterLayer.R
\name{is_RasterLayer}
\alias{is_RasterLayer}
\title{Test if RasterLayer}
\usage{
is_RasterLayer(x)
}
\arguments{
\item{x}{The object whose class is to be tested.}
}
\value{
logical; TRUE if class is of type \code{raster::RasterLayer}
}
\description{
Used to test if an object is of class \code{raster::RasterLayer}.
}
