% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Summarize HSI Models}
\usage{
model_summary(hsi_models, extent)
}
\arguments{
\item{hsi_models}{list; A named list of HSI model rasters.}

\item{extent}{raster::Extent; An extent object used to crop rasters.}
}
\value{
Returns a histogram and tmap for each of the HSI models.
}
\description{
Display a brief set of result summaries for each HSI model.
}
