% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_outside_model.R
\name{na_outside_model}
\alias{na_outside_model}
\title{Set NA Outside Modeled Area}
\usage{
na_outside_model(hsi_model)
}
\arguments{
\item{hsi_model}{raster; A raster containing HSI modeled values.}
}
\value{
A raster with unmodeled HSI cells originally set to -1, converted
to NA for summarization.
}
\description{
HSI values outside the modeled habitat zone were set to -1 to
distinguish areas inside the study area, but outside the modeled habitat
zone. This function is used for summarizing modeled habitat.
}
