% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_model.R
\name{name_model}
\alias{name_model}
\title{Name Raster HSI Model}
\usage{
name_model(hsi_model, model_name)
}
\arguments{
\item{hsi_model}{raster; A raster HSI model.}

\item{model_name}{character; The name of the model.}
}
\value{
A raster with the name assigned.
}
\description{
Assign a name to a raster HSI model. Used for processing
individual models into alternatives analysis for a study area.
}
