% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_fig.R
\name{raster_fig}
\alias{raster_fig}
\title{Create a Raster Figure}
\usage{
raster_fig(ras, extent, title = "Title")
}
\arguments{
\item{ras}{raster::RasterLayer; A raster object.}

\item{extent}{raster::Extent; An extent object used to crop the raster.}

\item{title}{character; A title for the raster legend.}
}
\value{
A tmap object
}
\description{
Creates a raster figure cropped to the specified extent
}
