% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_models.R
\name{recode_models}
\alias{recode_models}
\title{Recode HSI Models.}
\usage{
recode_models(model_list)
}
\arguments{
\item{model_list}{list; A named list of HSI model rasters.}
}
\value{
A named list of HSI models whose raster have had the -1 values
recoded to NA.
}
\description{
Recodes the -1 raster values used to denote the area within
the study area, but outside the model habitat zone to NA. HSI models
assign areas inside each model's habitat zone an HSI value of 0-1,
while areas outside a model's habitat zone are assigned a value of -1 to
distinguish it from areas outside the study area which are assigned the
value NoData. Some model summary functions require that these placeholder
values (-1) are converted to NA.
}
