% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raster_rdata.R
\name{save_raster_rdata}
\alias{save_raster_rdata}
\title{Save raster to RData file}
\usage{
save_raster_rdata(raster, file)
}
\arguments{
\item{raster}{RasterLayer; The raster layer to be saved.}

\item{file}{a (writable binary-mode) connection or the name of the file
where the data will be saved (when tilde expansion is done).}
}
\value{
None. Saves .RData file to disk.
}
\description{
Used to ensure that raster data is first loaded into memory so
that raster data is saved into the .RData object and not left in a
referenced temp file.
}
\details{
Using \code{save()} to write rasters to .RData files can produce
unexpected results. Large RasterLayers are not loaded into memory, but
contain a pointer to a file storing the data. \code{save()} then only stores a
pointer to the temp file holding the data which may not be available later.
This function loads on-disk RasterLayers into memory to ensure the data is
stored in the .RData file.
}
