% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_salinity_zone.R
\name{set_salinity_zone}
\alias{set_salinity_zone}
\title{Set Salinity Zone}
\usage{
set_salinity_zone(salinity, estuarine_marine_break, fresh_estuarine_break)
}
\arguments{
\item{salinity}{raster; A raster of salinity measurements or
estimates.}

\item{estuarine_marine_break}{numeric; The salinity threshold between
estuarine and marine zones.}

\item{fresh_estuarine_break}{numeric; The salinity threshold between
fresh and estuarine zones.}
}
\value{
A categorical salinity zone RasterLayer where:
\enumerate{
\item Marine
\item Estuarine
\item Fresh
}
}
\description{
Creates a Salinity Zone raster from the input salinity raster
applying the input threshold parameters.
}
