% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tidal_zone.R
\name{set_tidal_zone}
\alias{set_tidal_zone}
\title{Create a Tidal Zone Raster}
\usage{
set_tidal_zone(bed_elevation, MLLW, MHHW)
}
\arguments{
\item{bed_elevation}{RasterLayer; A bed elevation raster.}

\item{MLLW}{RasterLayer; A Mean Higher High Water (MHHW)
raster.}

\item{MHHW}{RasterLayer; A Mean Lower Low Water (MLLW)
raster.}
}
\value{
A categorical RasterLayer where:
\enumerate{
\item Deep
\item Subtidal
\item Intertidal
\item Upland
}
}
\description{
Creates a tidal zone raster from the input rasters.
}
\details{
The z-values of all input rasters must be expressed in the same
vertical datum (i.e., National Tidal Datum Epoch, NAVD88).
}
