% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by_poly.R
\name{summarize_by_poly}
\alias{summarize_by_poly}
\title{Summarize Raster by Polygon}
\usage{
summarize_by_poly(hsi_model, polys, progress = TRUE)
}
\arguments{
\item{hsi_model}{raster; A raster HSI model.}

\item{polys}{sf; A polygon used to derive be summaries from the
raster.}

\item{progress}{logical; TRUE displays a progress bar during processing.}
}
\value{
A data frame containing polygons summarized by the following fields:
\itemize{
\item ID - Polygon unique identifier. Derived from polys' row.names.
\item hu_<model_name> - HSI score mean for the polygon. Model name is derived
from the raster data names attribute via names().
\item count_<model_name> - Count of cells for the polygon.
\item acres_<model_name> - Area in acres for the polygon.
}
}
\description{
Summarize the raster HSI model by polygon.
}
