% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_models.R
\name{summarize_models}
\alias{summarize_models}
\title{Summarize HSI Models by Polygons}
\usage{
summarize_models(hsi_models, polys, progress = FALSE)
}
\arguments{
\item{hsi_models}{list; A named list of recoded HSI models created using
the \code{recode_models} function.}

\item{polys}{sf; A polygons sf object.}

\item{progress}{logical; TRUE displays a progress bar during processing.}
}
\value{
A data frame of model summaries containing the following fields
for each model:
\itemize{
\item ID - Polygon unique identifier. Derived from polys' row.names.
\item hu_<model_name> - HSI score mean for the polygon. Model name is derived
from the raster data names attribute via names().
\item count_<model_name> - Count of cells for the polygon.
\item acres_<model_name> - Area in acres for the polygon.
}
}
\description{
Summarizes the list of models by each polygon. This function
makes a call to the \code{summarize_by_poly} function for each model in the
\code{hsi_models} list. This function assembles those results into a single
output data frame.
}
