% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{HSI Model Summary Table.}
\usage{
summary_table(
  summary_df,
  model_metric = c("hsi_", "acres_", "hu_"),
  polys,
  poly_field,
  caption,
  digits = 2
)
}
\arguments{
\item{summary_df}{data.frame; A model summary data frame created by the
\code{summarize_models} function.}

\item{model_metric}{character; The model metric to be summarized from the
\code{summary_df}. One of "hsi_", "acres_", or "hu_".}

\item{polys}{sf; A sf polygon object.}

\item{poly_field}{character; The \code{polys} field used to identify the
summarization features.}

\item{caption}{character; A caption for the table.}

\item{digits}{numeric; Number of decimal places to round metric.}
}
\value{
A model summary data frame and a \code{knitr::kable} table object (is
returned silently).
}
\description{
Produces an HSI model summary table for the specified metric.
}
