% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cat_tiff.R
\name{write_cat_tif}
\alias{write_cat_tif}
\title{Write Categorical Raster TIFF file}
\usage{
write_cat_tif(cat_raster, out_dir, out_name)
}
\arguments{
\item{cat_raster}{raster::Raster*; A categorical raster.}

\item{out_dir}{character; A path to a folder.}

\item{out_name}{character; A file name.}
}
\value{
This function is used for the side-effect of writing a categorical
raster to a file.
}
\description{
Writes a categorical raster file in .tiff format with the
correct raster attribute table information and compression to be read by
ESRI ArcGIS products.
}
