% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cont_tiff.R
\name{write_cont_tiff}
\alias{write_cont_tiff}
\title{Write Continuous Raster TIFF file}
\usage{
write_cont_tiff(cont_raster, out_dir, out_name)
}
\arguments{
\item{cont_raster}{raster::Raster*; A continuous raster.}

\item{out_dir}{character; A path to a folder.}

\item{out_name}{character; A file name.}
}
\value{
This function is used for the side-effect of writing a continuous
raster to a file.
}
\description{
Writes a continuous raster file in .tiff format with the
correct raster attribute table information to be read by ESRI ArcGIS
products.
}
