% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSIcalc.R
\name{HSIcalc}
\alias{HSIcalc}
\title{Calculate Habitat Suitability Index (HSI)}
\usage{
HSIcalc(si_list, method = "mean")
}
\arguments{
\item{si_list}{list; A list of suitability indices.}

\item{method}{character; The method for calculating the suitability
index. Must be one of "mean",}
}
\value{
An HSI object matching the data format of the input suitability
indices.
}
\description{
Calculates a habitat suitability index for the input
suitability indices using the specified method.
}
\examples{
# Build manual model
# Number of trees > 51cm diameter per 0.4 ha plot
tree.num     <- c(0, 2, 4, NA)              # parameter breakpoints
tree.num.SIV <- c(0.1, 1, 1, NA)            # parameter suitability indices
tree_num <- data.frame(tree.num, tree.num.SIV)

# Mean diameter of overstory trees
avg.dbh      <- c(0, 5, 20, NA)             # parameter breakpoints
avg.dbh.SIV  <- c(0, 0, 1, NA)              # parameter suitability indices
tree_diameter <- data.frame(avg.dbh, avg.dbh.SIV)

# Percent canopy cover of overstory trees
can.cov      <- c(0, 20, 60, 100)           # parameter breakpoints
can.cov.SIV  <- c(0, 0, 1, 1)               # parameter suitability indices
canopy_cov <- data.frame(can.cov, can.cov.SIV)

barredowl <- data.frame(tree_num, tree_diameter, canopy_cov)

bo_test1 <- list(2, 5, 20)

# Barredowl si values
bo_si_1 <- nybem::SIcalc(barredowl, bo_test1)

# Calculate HSI
bo_hsi <- HSIcalc(si_list = bo_si_1)

}
