% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nybem_submodels.R
\docType{data}
\name{nybem_submodels}
\alias{nybem_submodels}
\title{NYBEM Submodels}
\format{
A list with the following named submodels:
\itemize{
\item NYBEM.fresh.tid - Freshwater, Tidal Zone model.
\item NYBEM.est.int - Estuarine, Intertidal Zone model.
\item NYBEM.est.sub.hard - Estuarine, Subtidal Zone, Hard Bottom model.
\item NYBEM.est.sub.soft.sav - Estuarine, Subtidal Zone, Soft Bottom, Submersed
Aquatic Vegetation model.
\item NYBEM.est.sub.soft.clam - Estuarine, Subtidal Zone, Soft Bottom, Hard
Clam model.
\item NYBEM.mar.int - Marine, Intertidal Zone model.
\item NYBEM.mar.sub - Marine, Subtidal Zone model.
\item NYBEM.mar.deep - Marine, Deepwater Zone model.
}
}
\usage{
nybem_submodels
}
\description{
A list of submodels used in the NYBEM feasibility study.
}
\keyword{datasets}
