% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by_poly.R
\name{summarize_by_poly}
\alias{summarize_by_poly}
\title{Summarize Raster by Polygon}
\usage{
summarize_by_poly(hsi_model, polys)
}
\arguments{
\item{hsi_model}{raster; A raster HSI model.}

\item{polys}{sf; A polygon used to derive be summaries from the
raster.}
}
\value{
A data frame containing polygons summarized by the following fields:
\itemize{
\item hu_<model_name> - HSI score mean for the polygon.
\item acres_<model_name> - Area in acres for the polygon.
}
}
\description{
Summarize the raster HSI model by polygon.
}
